requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
Pillow==8.2.0
streamlit==1.23.1